% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAllPairsDistances.R
\name{calcAllPairsDistances}
\alias{calcAllPairsDistances}
\title{Pre-processing Step}
\usage{
calcAllPairsDistances(network, networkName, directionPaths = "all",
  weightVector = NULL)
}
\arguments{
\item{network}{An igraph network on which to calculate the distances.}

\item{networkName}{A string containing the name of the network. Required.}

\item{directionPaths}{defaults to "all". A string indicating how the
distances should be calculated.}

\item{weightVector}{defaults to NULL. A numeric vector giving edge weights.}
}
\description{
Before applying GeneSurrounder, the distances on the 
global network should be calculated
}
