% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_cores}
\alias{check_cores}
\title{Check number of cores}
\usage{
check_cores(numCores)
}
\arguments{
\item{numCores}{the number of cores}
}
\value{
the corrected number of cores
}
\description{
This internal function checks if the number of cores
specified is reasonable.
}
