% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_resample_size}
\alias{check_resample_size}
\title{Check Resample Size}
\usage{
check_resample_size(classLabels, numResamples, allPerms)
}
\arguments{
\item{classLabels}{the vector of labels for the 
experiment to determine maximum possible permutations}

\item{numResamples}{the number of resamples}

\item{allPerms}{boolean for whether all permutations should be
generated}
}
\value{
a list with the corrected numResamples and allPerms
}
\description{
This internal function checks to make sure
the specified number of resamples is correct.
If it was not specified, then it is set as the
maximum number of permutations.
}
