% Generated by roxygen2: do not edit by hand
% Please edit documentation in 2016-04-06/geneNIDG.R
\name{geneNIDG}
\alias{geneNIDG}
\title{GeneSurrounder wrapper}
\usage{
geneNIDG(distance.matrix, cor.matrix, geneStats, genes.assayedETnetwork,
  diameter, num.resamples, gene.id)
}
\arguments{
\item{distance.matrix}{This is a description.}

\item{cor.matrix}{This is a description.}

\item{geneStats}{This is a description.}

\item{genes.assayedETnetwork}{This is a description.}

\item{diameter}{This is a description.}

\item{num.resamples}{This is a description.}

\item{gene.id}{This is a description.}
}
\description{
The GeneSurrounder method consists of two tests that are run 
independently of each other and then combined to determine if 
the putative disease gene is a “disruptive” candidate disease gene 
meeting both criteria.
}

