% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_sleuth_stats}
\alias{get_sleuth_stats}
\title{Get sleuth Stats}
\usage{
get_sleuth_stats(obj, testType, whichModel, whichTest)
}
\arguments{
\item{obj}{the sleuth object. It must have 'whichModel' already fit
using \code{sleuth_fit}, as well as the null model if the testType is
'lrt'.}

\item{testType}{either "lrt" or "wt" to indicate which test, Likelihood Ratio Test or
Wald Test, to use to calculate statistics}

\item{whichModel}{if "lrt" is the test type, this the alternative model; if "wt" is
the test type, this is the model to use.}

\item{whichTest}{if "lrt", this is the test in the format of "[null model]:[alternative model]";
if "wt", this is the same as whichBeta, the actual column in the design matrix to test}
}
\value{
a vector with the statistics from the specified test.
}
\description{
This internal function retrieves
the specified test statistics from
the sleuth object.
}
