% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Observed.DecayDE.R
\name{Observed.DecayDE}
\alias{Observed.DecayDE}
\title{Decay of Differential Expression step}
\usage{
Observed.DecayDE(distance_matrix, gene_id, genes_assayedETnetwork, diameter,
  geneStats_observed)
}
\arguments{
\item{distance_matrix}{A matrix of the distances on the global network}

\item{gene_id}{The name of the gene to which GeneSurrounder is applied}

\item{genes_assayedETnetwork}{The names of the genes that are assayed and on the network}

\item{diameter}{The diameter of the global network}

\item{geneStats_observed}{A vector of the observed differential expression}
}
\description{
Decay of Differential Expression, tests whether the magnitude of 
differential expression of other genes j in the neighborhood is inversely 
related to the distance d(i,j) of gene j from gene i
}
