% Generated by roxygen2: do not edit by hand
% Please edit documentation in 2016-01-14/Resample.SI.R
\name{Resample.SI}
\alias{Resample.SI}
\title{Sphere of Influence Step}
\usage{
Resample.SI(gene.id, distance.matrix, cor.matrix, diameter,
  num.resamples = 1000, genes.assayedETnetwork)
}
\arguments{
\item{gene.id}{This is a description.}

\item{distance.matrix}{This is a description.}

\item{cor.matrix}{This is a description.}

\item{diameter}{This is a description.}

\item{num.resamples}{default 1000.}

\item{genes.assayedETnetwork}{This is a description.}
}
\description{
Sphere of Influence assesses if a 
candidate gene i meets the first criterion by testing if gene i 
is more strongly correlated with its network neighbors than with a 
random set of genes
}

