% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{returnResampleMat}
\alias{returnResampleMat}
\title{Return Matrix of Unique Resamplings}
\usage{
returnResampleMat(classLabels, numResamples = NULL, allPerms = TRUE)
}
\arguments{
\item{classLabels}{a named vector with the class labels to be permuted.
at this time, only two-condition experiments are allowed.}

\item{numResamples}{the number of desired unique resamplings;
if this value is equal to or greater than the maximum possible
unique resamplings, all unique resamplings will be provided.
The default is \code{NULL}, indicating that all combinations should be
generated.}

\item{allPerms}{if \code{TRUE} (the default), it will override \code{numResamples}
and generate all possible unique combinations.}
}
\value{
a matrix with n rows by m columns, with n equal to the number of
  resamplings, and m equal to the length of the class labels.
}
\description{
This method takes a named vector of class labels, and
returns a matrix of the resampled class labels matched to
the original sample names. This only generates unique
combinations (i.e. ordering does not matter).
}
