% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SumAbsCor.R
\name{SumAbsCor}
\alias{SumAbsCor}
\title{Helper function for Sphere of Influence procedure}
\usage{
SumAbsCor(cor.vector, diameter, distances.to.j)
}
\arguments{
\item{cor.vector}{A vector of correlations}

\item{diameter}{The diameter of the global network}

\item{distances.to.j}{A vector of distances to the gene to which GeneSurrounder is applied}
}
\description{
\code{SumAbsCor} returns the total observed correlation between gene i and its neighbors
}

