% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/T2gUtils.R
\name{transToGeneCors}
\alias{transToGeneCors}
\title{Aggregate transcript-level correlations to gene-level}
\usage{
transToGeneCors(data, t2g, numCores = 1L)
}
\arguments{
\item{data}{a matrix of transcript expression values. The rows
should be samples and the columns should be transcripts,
with the column names being the transcript IDs used to map
to the gene level.}

\item{t2g}{a data frame with two columns; the first contains
the transcript IDs, and should match the names of pvals; the second
contains the gene IDs. All transcripts should have one
and only one corresponding gene ID.}

\item{numCores}{the number of cores to use for parallel computation}
}
\value{
an 'ff' object pointing to a file on disk containing the
  estimated gene-level correlation matrix
}
\description{
This method uses the \code{whitening} package to
conduct a canonical correlation analysis between each
pair of genes, using each gene's transcripts as the
set of variables under consideration, and using the shrinkage
estimate from the \code{scca} function within \code{whitening}.
This allows for estimation of a gene-level correlation
by examining the canonical correlations between the two sets
of transcripts.
}
