% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/T2gUtils.R
\name{transToGenePvals}
\alias{transToGenePvals}
\title{Aggregate transcript-level p-values to gene-level}
\usage{
transToGenePvals(pvals, mean_obs, t2g, weight_func = identity, numCores = 1)
}
\arguments{
\item{pvals}{vector or matrix of transcript-level p-values.
if a vector, names should be transcript IDs; if a matrix,
rownames should be transcript IDs.}

\item{mean_obs}{named vector of the mean of the observed normalized
expression for each transcript}

\item{t2g}{a data frame with two columns; the first contains
the transcript IDs, and should match the names of pvals; the second
contains the gene IDs. All transcripts should have one
and only one corresponding gene ID.}

\item{weight_func}{which function should be applied to mean_obs
to weight the p-values for aggregation using the lancaster method?
the default is the identity function. Note that whatever
function is applied should yield only positive values.}

\item{numCores}{the number of cores to use. If a p-value matrix is supplied
and numCores is set to 1, this function will use data.table to calculate
the matrix gene-level matrix. If it's more than one, it will use
parallel's mclapply, which is faster based on benchmarks.}
}
\value{
a vector or matrix of gene-level p-values. The names of
  the vector, or the rownames of the matrix, will be the provided
  gene IDs.
}
\description{
This takes a vector or a matrix of transcript-level
p-values and aggregates them to the gene-level.
}
