% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCorMatrix.R
\name{calcCorMatrix}
\alias{calcCorMatrix}
\title{Pre-processing Step}
\usage{
calcCorMatrix(exprMatrix, corMethod, exprName)
}
\arguments{
\item{exprMatrix}{A matrix (genes by samples) of expression values.}

\item{corMethod}{A string contating the correlation method to use.}

\item{exprName}{A string containg the name of the expression matrix.}
}
\description{
Before applying GeneSurrounder, the correlation
between the expression of the genes should be calculated
}

