% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/T2gUtils.R
\name{calcGeneCor}
\alias{calcGeneCor}
\title{Calculate large gene expression correlation matrix in parallel}
\usage{
calcGeneCor(data, size = 2000, numCores = 1L)
}
\arguments{
\item{data}{a N x M matrix or data.frame, with N columns corresponding to
each gene observed, and M rows corresponding to the samples
in the experiment.}

\item{size}{the number of elements per block to split the data. The
default is 2000, which results in decent performance for typical
usage (order of 10-20K features). If the number of features is
not evenly divisible by the size parameter, a remainder block is used.}

\item{numCores}{the number of cores to use for parallel calculations}
}
\value{
an 'ff' object pointing to a file on disk containing the
  gene correlation matrix
}
\description{
This methods uses the ff package to calculate a large
correlation matrix.
}
