% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcGeneTStats.R
\name{calcGeneTStats}
\alias{calcGeneTStats}
\title{Pre-processing Step}
\usage{
calcGeneTStats(expr, classLabels, numResamples = 1000)
}
\arguments{
\item{expr}{A matrix (genes by samples) of expression values.}

\item{classLabels}{A factor with levels corresponding to class labels.}

\item{numResamples}{defaults to 1000. The number of resamples when calculating resampled differential expression.}
}
\description{
Before applying GeneSurrounder, the observed and resampled 
differential expression of the genes should be calculated
}

