% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneNIDG.R
\name{geneNIDG}
\alias{geneNIDG}
\title{GeneSurrounder wrapper}
\usage{
geneNIDG(distance_matrix, cor_matrix, geneStats_observed, perm_geneStats_matrix,
  genes_assayedETnetwork, diameter, num_Sphere_resamples, gene_id)
}
\arguments{
\item{distance_matrix}{A matrix of the distances on the global network}

\item{cor_matrix}{A matrix of correlations between the expression of the genes}

\item{geneStats_observed}{A vector of the observed differential expression}

\item{perm_geneStats_matrix}{A matrix of the resampled differential expression}

\item{genes_assayedETnetwork}{The names of the genes that are assayed and on the network}

\item{diameter}{The diameter of the global network}

\item{num_Sphere_resamples}{The number of resamples when running the Sphere of Influence Procedure}

\item{gene_id}{The name of the gene to which GeneSurrounder is applied}
}
\description{
The GeneSurrounder method consists of two tests that are run 
independently of each other and then combined to determine if 
the putative disease gene is a “disruptive” candidate disease gene 
meeting both criteria.
}
