% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Observed.SI.R
\name{Observed.SI}
\alias{Observed.SI}
\title{Sphere of Influence Step}
\usage{
Observed.SI(gene_id, distance_matrix, cor_matrix, diameter,
  genes_assayedETnetwork)
}
\arguments{
\item{gene_id}{The name of the gene to which GeneSurrounder is applied}

\item{distance_matrix}{A matrix of the distances on the global network}

\item{cor_matrix}{A matrix of correlations between the expression of the genes}

\item{diameter}{The diameter of the global network}

\item{genes_assayedETnetwork}{The names of the genes that are assayed and on the network}
}
\description{
Sphere of Influence assesses if a 
candidate gene i meets the first criterion by testing if gene i 
is more strongly correlated with its network neighbors than with a 
random set of genes
}
