% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRadiusVS.R
\name{plotRadiusVS}
\alias{plotRadiusVS}
\title{Plot results of GeneSurrounder}
\usage{
plotRadiusVS(geneNIDG.gene)
}
\arguments{
\item{geneNIDG.gene}{A dataframe that is the output of the geneNIDG function}
}
\description{
\code{plotRadiusVS} plots the results of GeneSurrounder using the output of the geneNIDG function
}
