% Generated by roxygen2: do not edit by hand
% Please edit documentation in 2016-04-06/Resample.DecayDE.R
\name{Resample.DecayDE}
\alias{Resample.DecayDE}
\title{Decay of Differential Expression step}
\usage{
Resample.DecayDE(distance.matrix, gene.id, genes.assayedETnetwork, diameter,
  geneStats, sizes)
}
\arguments{
\item{distance.matrix}{This is a description.}

\item{gene.id}{This is a description.}

\item{genes.assayedETnetwork}{This is a description.}

\item{diameter}{This is a description.}

\item{geneStats}{This is a description.}

\item{sizes}{This is a description.}
}
\description{
Decay of Differential Expression tests whether the magnitude of 
differential expression of other genes j in the neighborhood is inversely 
related to the distance d(i,j) of gene j from gene i
}

