% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Resample.SI.R
\name{Resample.SI}
\alias{Resample.SI}
\title{Sphere of Influence Step}
\usage{
Resample.SI(gene.id, distance.matrix, cor.matrix, diameter,
  num.Sphere.resamples = 1000, genes.assayedETnetwork)
}
\arguments{
\item{gene.id}{The name of the gene to which GeneSurrounder is applied}

\item{distance.matrix}{A matrix of the distances on the global network}

\item{cor.matrix}{A matrix of correlations between the expression of the genes}

\item{diameter}{The diameter of the global network}

\item{num.Sphere.resamples}{defaults to 1000. The number of resamples when running the Sphere of Influence Procedure}

\item{genes.assayedETnetwork}{The names of the genes that are assayed and on the network}
}
\description{
Sphere of Influence assesses if a 
candidate gene i meets the first criterion by testing if gene i 
is more strongly correlated with its network neighbors than with a 
random set of genes
}

