% Generated by roxygen2: do not edit by hand
% Please edit documentation in 2016-01-14/SumAbsCor.R
\name{SumAbsCor}
\alias{SumAbsCor}
\title{Sum of vector elements.}
\usage{
SumAbsCor(cor.vector, diameter, distances.to.j)
}
\arguments{
\item{cor.vector}{This is a description.}

\item{diameter}{This is a description.}

\item{distances.to.j}{This is a description.}
}
\description{
\code{SumAbsCor} returns the total observed correlation between gene i and its neighbors
}

