% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SumAbsCor.R
\name{SumAbsCor}
\alias{SumAbsCor}
\title{Helper function for Sphere of Influence procedure}
\usage{
SumAbsCor(cor_vector, diameter, distances_to_j)
}
\arguments{
\item{cor_vector}{A vector of correlations}

\item{diameter}{The diameter of the global network}

\item{distances_to_j}{A vector of distances to the gene to which GeneSurrounder is applied}
}
\description{
\code{SumAbsCor} returns the total observed correlation between gene i and its neighbors
}
